Refactoring: moved outcomes related interfaces into a separate model
