+++ src/models/Outcomes.ts
@@ -1,0 +1,13 @@
+export interface OutcomesConfig {
+  direct: {
+    enabled: boolean;
+  }
+  indirect: {
+    enabled: boolean;
+    influencedTimePeriodMin: number;
+    influencedNotificationsLimit: number;
+  }
+  unattributed: {
+    enabled: boolean;
+  }
+}
